+++ TTKModule/TTKCore/musicNetworkKits/kw/musicdownloadquerykwthread.cpp
@@ -72,6 +72,7 @@ MusicDownLoadQueryKWThread::MusicDownLoadQueryKWThread(QObject *parent)
     : MusicDownLoadQueryThreadAbstract(parent)
 {
     m_queryServer = "Kuwo";
+    m_pageSize = 40;
 }
 
 QString MusicDownLoadQueryKWThread::getClassName()
@@ -87,11 +88,30 @@ void MusicDownLoadQueryKWThread::startToSearch(QueryType type, const QString &te
     }
 
     M_LOGGER_INFO(QString("%1 startToSearch %2").arg(getClassName()).arg(text));
-    m_searchText = text.trimmed();
     m_currentType = type;
-    QUrl musicUrl = MusicUtils::Algorithm::mdII(KW_SONG_SEARCH_URL, false).arg(text).arg(0).arg(50);
+    m_searchText = text.trimmed();
+
+    emit clearAllItems();
+    m_musicSongInfos.clear();
+
+    startToPage(0);
+}
+
+void MusicDownLoadQueryKWThread::startToPage(int offset)
+{
+    if(!m_manager)
+    {
+        return;
+    }
+
+    M_LOGGER_INFO(QString("%1 startToPage %2").arg(getClassName()).arg(offset));
     deleteAll();
+
+    QUrl musicUrl = MusicUtils::Algorithm::mdII(KW_SONG_SEARCH_URL, false)
+                    .arg(m_searchText).arg(offset).arg(m_pageSize);
     m_interrupt = true;
+    m_pageTotal = 0;
+    m_pageIndex = offset;
 
     QNetworkRequest request;
     request.setUrl(musicUrl);
@@ -136,8 +156,6 @@ void MusicDownLoadQueryKWThread::downLoadFinished()
     }
 
     M_LOGGER_INFO(QString("%1 downLoadFinished").arg(getClassName()));
-    emit clearAllItems();
-    m_musicSongInfos.clear();
     m_interrupt = false;
 
     if(m_reply->error() == QNetworkReply::NoError)
@@ -153,6 +171,7 @@ void MusicDownLoadQueryKWThread::downLoadFinished()
             QVariantMap value = data.toMap();
             if(value.contains("abslist"))
             {
+                m_pageTotal = value["TOTAL"].toInt();
                 QVariantList datas = value["abslist"].toList();
                 foreach(const QVariant &var, datas)
                 {
